update libraries move categories mocks folder
