+++ .gitignore
@@ -5,6 +5,7 @@
 *.so
 *.dylib
 .vscode
+__debug_bin
 
 # Test binary, built with `go test -c`
 *.test
