+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM golang:1.14 as builder
+FROM golang:1.15 as builder
 WORKDIR /
 COPY . ./
 RUN CGO_ENABLED=0 GOOS=linux go build -a -installsuffix .
