+++ categories/mocks/entities.go
