+++ go.mod
@@ -5,11 +5,11 @@ go 1.15
 require (
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
 	github.com/gin-gonic/gin v1.6.3
-	github.com/go-playground/validator/v10 v10.3.0
+	github.com/go-playground/validator/v10 v10.4.1
 	github.com/google/uuid v1.1.2
 	github.com/pkg/errors v0.9.1
 	github.com/stretchr/testify v1.6.1
-	golang.org/x/crypto v0.0.0-20200820211705-5c72a883971a // indirect
-	gorm.io/driver/postgres v1.0.1
-	gorm.io/gorm v1.20.2
+	golang.org/x/crypto v0.0.0-20201221181555-eec23a3978ad // indirect
+	gorm.io/driver/postgres v1.0.6
+	gorm.io/gorm v1.20.9
 )
